+++ src/main/java/com/example/crm_project/service/TransactionService.java
@@ -1,12 +1,13 @@
 package com.example.crm_project.service;
 
+import com.example.crm_project.exception.TransactionNotFoundException;
 import com.example.crm_project.model.Transaction;
 import com.example.crm_project.repository.TransactionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+
 
 @Service
 @RequiredArgsConstructor
@@ -18,8 +19,9 @@ public class TransactionService {
         return transactionRepository.findAll();
     }
 
-    public Optional<Transaction> findById(Long id) {
-        return transactionRepository.findById(id);
+    public Transaction findById(Long id) {
+        return transactionRepository.findById(id)
+                .orElseThrow(() -> new TransactionNotFoundException("Транзакция с id " + id + " не найдена"));
     }
 
     public Transaction save(Transaction transaction) {
@@ -27,6 +29,9 @@ public class TransactionService {
     }
 
     public void deleteById(Long id) {
+        if (!transactionRepository.existsById(id)) {
+            throw new TransactionNotFoundException("Транзакция с id " + id + " не найдена");
+        }
         transactionRepository.deleteById(id);
     }
